refactor(core): LearnCore 도메인 타입 통합
병렬 4세션 병합으로 LearnCore 안에 같은 개념이 두 벌 생겼던 것을 한 벌로
합치고, 공유 도메인 타입을 새 디렉터리 Sources/LearnCore/Review/ 로 옮겼다.

- ReviewRating / CardPhase / ReviewLogSource / EpochMillis / ReviewLogEntry /
  SchedulerParameterSet / CardSchedulingState / CardStateSnapshot → Review/
- Persistence/ 에는 저장 고유의 것만, Scheduling/ 에는 스케줄러 고유의 것만 남겼다
- 시각은 EpochMillis(struct) 로 통일하고 typealias EpochMilliseconds 를 제거
- 학습 단계는 CardPhase(Int) 하나. DB 는 그대로 TEXT 이고 정수↔TEXT 변환은
  LearnPersistence/Rows/DomainColumns.swift 가 전담한다 (골든 스키마 불변)
- ReviewLogSource 는 review/cram/manual/imported 합집합. 출시된 CHECK 리터럴
  ('scheduled'·'import') 스펠링도 같은 매핑 계층이 흡수한다
- ReviewLogEntry 는 newtype id(ReviewLogID·ParameterSetID) + affectsSchedule +
  replayOrdered() 3중 정렬 키를 모두 보존
- CardStateSnapshot 이 CardSchedulingState 를 합성하도록 정리 (필드 중복 제거)
- Package.swift 경고 2건 제거, 스케줄링 픽스처는 Bundle.module 로 읽는다
- GradeResult.Presenter 에 CaseIterable 추가, 관련 테스트를 allCases 기반으로
- RunFailure.cpuExceeded 를 다루지 않던 스위치 2곳 보강 (병합 시 누락분)

swift build 경고 0, swift test 316개 전부 통과.

+++ Packages/LearnKit/Tests/LearnPersistenceTests/ConcurrencyBoundaryTests.swift
@@ -98,7 +98,7 @@ struct ConcurrencyBoundaryTests {
             for index in 0..<40 {
                 group.addTask {
                     try await store.append(
-                        Fixture.reviewEntry(card: "card-\(index % 4)", at: Int64(index))
+                        Fixture.reviewEntry(card: "card-\(index % 4)", at: index)
                     )
                 }
             }
@@ -119,7 +119,7 @@ struct ConcurrencyBoundaryTests {
 
         try await withThrowingTaskGroup(of: Void.self) { group in
             for index in 0..<20 {
-                group.addTask { try await log.append(Fixture.reviewEntry(at: Int64(index))) }
+                group.addTask { try await log.append(Fixture.reviewEntry(at: index)) }
                 group.addTask { _ = try await cards.dueCards(
                     languageID: .python, dueAtOrBefore: .max, limit: 10
                 ) }
